refactor: 클래스명 변경 및 위치 변경
