+++ app/src/main/java/woowacourse/shopping/data/cart/CartRepository.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.data.cart
+
+import woowacourse.shopping.model.product.Product
+
+interface CartRepository {
+    val products: MutableList<Product>
+
+    fun add(product: Product)
+
+    fun remove(product: Product)
+}
