+++ app/src/main/java/woowacourse/shopping/data/products/Products.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping.model.products
+package woowacourse.shopping.data.products
+
+import woowacourse.shopping.model.product.Product
 
 val dummyProductsData =
     listOf(
