+++ app/src/main/java/woowacourse/shopping/model/product/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.products
+package woowacourse.shopping.model.product
 
 import java.io.Serializable
