+++ app/src/main/res/layout/activity_product_detail.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.model.products.Product" />
+            type="woowacourse.shopping.model.product.Product" />
 
         <variable
             name="viewModel"
