+++ app/src/main/res/layout/item_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.model.products.Product" />
+            type="woowacourse.shopping.model.product.Product" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
